--- backend/real_estate_rental_system/src/main/java/cs425group3_real_estate_rental_system/RealEstateRentalSystemApplication.java
@@ -1,13 +1,0 @@
-package cs425group3_real_estate_rental_system;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RealEstateRentalSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RealEstateRentalSystemApplication.class, args);
-	}
-
-}
