+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/Booking/BookingStatus.java
@@ -1,4 +1,4 @@
-package cs425group3_real_estate_rental_system.Booking;
+package group3_real_estate_rental_system.Booking;
 
 public enum BookingStatus {
