+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/Lease/Lease.java
@@ -1,23 +1,23 @@
-package cs425group3_real_estate_rental_system.Lease;
+package group3_real_estate_rental_system.Lease;
 
-import cs425group3_real_estate_rental_system.Property.Property;
-import cs425group3_real_estate_rental_system.User.entity.PropertyOwner;
-import cs425group3_real_estate_rental_system.User.entity.Tenant;
+import group3_real_estate_rental_system.Property.Property;
+import group3_real_estate_rental_system.User.entity.PropertyOwner;
+import group3_real_estate_rental_system.User.entity.Tenant;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor
-@RequiredArgsConstructor
+@AllArgsConstructor
 @Getter
 @ToString
 @EqualsAndHashCode
 public class Lease {
     @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long LeaseId;
+
     private String leaseURL;
     private LocalDateTime tenantSignedDate;
     private LocalDateTime propertyOwnerSignedDate;
