+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/User/entity/Admin.java
@@ -1,20 +1,18 @@
-package cs425group3_real_estate_rental_system.User.entity;
+package group3_real_estate_rental_system.User.entity;
 
-import cs425group3_real_estate_rental_system.Lease.Lease;
+import group3_real_estate_rental_system.Lease.Lease;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.List;
 
 @Entity
-@RequiredArgsConstructor
-@NoArgsConstructor
 @Getter
 @Setter
 @ToString
 @EqualsAndHashCode
 public class Admin extends User {
-    private Long adminId;
+    private String role;
     @OneToMany(mappedBy = "propertyOwner", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Lease> leases;
