+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/User/entity/PropertyOwner.java
@@ -1,9 +1,8 @@
-package cs425group3_real_estate_rental_system.User.entity;
+package group3_real_estate_rental_system.User.entity;
 
-import cs425group3_real_estate_rental_system.Booking.Booking;
-import cs425group3_real_estate_rental_system.Lease.Lease;
-import cs425group3_real_estate_rental_system.Property.Property;
-import cs425group3_real_estate_rental_system.User.entity.User;
+import group3_real_estate_rental_system.Booking.Booking;
+import group3_real_estate_rental_system.Lease.Lease;
+import group3_real_estate_rental_system.Property.Property;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -11,8 +10,6 @@ import lombok.*;
 import java.util.List;
 
 @Entity
-@RequiredArgsConstructor
-@NoArgsConstructor
 @Getter
 @Setter
 @ToString
@@ -20,10 +17,12 @@ import java.util.List;
 public class PropertyOwner extends User {
 
     private Long propertyOwnerId;
-    @OneToMany(mappedBy = "owner", cascade = CascadeType.ALL, orphanRemoval = true)
+    @OneToMany(mappedBy = "propertyOwner", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Lease> leases;
+
     @OneToMany(mappedBy = "owner", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Property> properties;
+
     @OneToMany(mappedBy = "approvedBy", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Booking> Bookings;
