+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/User/entity/Tenant.java
@@ -1,20 +1,19 @@
-package cs425group3_real_estate_rental_system.User.entity;
+package group3_real_estate_rental_system.User.entity;
 
-import cs425group3_real_estate_rental_system.Booking.Booking;
-import cs425group3_real_estate_rental_system.Lease.Lease;
+import group3_real_estate_rental_system.Booking.Booking;
+import group3_real_estate_rental_system.Lease.Lease;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.List;
 
 @Entity
-@RequiredArgsConstructor
-@NoArgsConstructor
 @Getter
 @Setter
 @ToString
 @EqualsAndHashCode
 public class Tenant extends User {
+
     private Long tenantId;
     @OneToMany(mappedBy = "tenant", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Lease> leases;
