+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/User/entity/User.java
@@ -1,15 +1,20 @@
-package cs425group3_real_estate_rental_system.User.entity;
+package group3_real_estate_rental_system.User.entity;
 
-import cs425group3_real_estate_rental_system.common.Address;
+import group3_real_estate_rental_system.common.Address;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
+@AllArgsConstructor
+@NoArgsConstructor
 @Data
 @Inheritance(strategy = InheritanceType.JOINED)
 public class User {
     @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long userId;
+
     private String firstName;
     private String lastName;
     private String username;
