+++ backend/real_estate_rental_system/src/main/java/group3_real_estate_rental_system/common/Address.java
@@ -1,0 +1,15 @@
+package group3_real_estate_rental_system.common;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public record Address (
+        String street,
+        String city,
+        String state,
+        String zipCode,
+        String country
+){}
+
+
+
